refactor: readings routing and components
