+++ src/components/ReadingComponents/ButtonAgregar.js
@@ -5,7 +5,7 @@ import { useNavigate } from "react-router-dom";
 const AgregarButton = ({ onClick, userId }) => {
     const navigate = useNavigate();
     return(
-        <button className="btn btn-primary" onClick={() => navigate("/add-genre", { state: { userId } })}>
+        <button className="btn btn-primary" onClick={() => navigate("/readings-list/add-genre", { state: { userId } })}>
             Agregar
         </button>
     )
